feat: remove IRenderAuxGeom
