+++ Code/Framework/AzFramework/AzFramework/Entity/EntityDebugDisplayBus.h
@@ -33,6 +33,156 @@ namespace AzFramework
 {
     inline constexpr AZ::s32 g_defaultSceneEntityDebugDisplayId = AZ_CRC_CE("MainViewportEntityDebugDisplayId"); // default id to draw to all viewports in the default scene
 
+    // Notes:
+    //   Don't change the xxxShift values, they need to match the values from legacy cry rendering
+    //   This also applies to the individual flags in EAuxGeomPublicRenderflags_*!
+    // Remarks:
+    //   Bits 0 - 22 are currently reserved for prim type and per draw call render parameters (point size, etc.)
+    //   Check RenderAuxGeom.h in ../RenderDll/Common
+    enum EAuxGeomPublicRenderflagBitMasks
+    {
+        e_Mode2D3DShift             = 31,
+        e_Mode2D3DMask              = 0x1 << e_Mode2D3DShift,
+
+        e_AlphaBlendingShift        = 29,
+        e_AlphaBlendingMask         = 0x3 << e_AlphaBlendingShift,
+
+        e_DrawInFrontShift          = 28,
+        e_DrawInFrontMask           = 0x1 << e_DrawInFrontShift,
+
+        e_FillModeShift             = 26,
+        e_FillModeMask              = 0x3 << e_FillModeShift,
+
+        e_CullModeShift             = 24,
+        e_CullModeMask              = 0x3 << e_CullModeShift,
+
+        e_DepthWriteShift           = 23,
+        e_DepthWriteMask            = 0x1 << e_DepthWriteShift,
+
+        e_DepthTestShift            = 22,
+        e_DepthTestMask             = 0x1 << e_DepthTestShift,
+
+        e_PublicParamsMask      = e_Mode2D3DMask | e_AlphaBlendingMask | e_DrawInFrontMask | e_FillModeMask |
+        e_CullModeMask | e_DepthWriteMask | e_DepthTestMask
+    };
+
+    // Notes:
+    //   e_Mode2D renders in normalized [0.. 1] screen space.
+    //   Don't change the xxxShift values blindly as they affect the rendering output
+    //   that is two primitives have to be rendered after 3d primitives, alpha blended
+    //   geometry have to be rendered after opaque ones, etc.
+    //   This also applies to the individual flags in EAuxGeomPublicRenderflagBitMasks!
+    // Remarks:
+    //   Bits 0 - 22 are currently reserved for prim type and per draw call render parameters (point size, etc.)
+    //   Check RenderAuxGeom.h in ../RenderDll/Common
+    // See also:
+    //   EAuxGeomPublicRenderflagBitMasks
+    enum EAuxGeomPublicRenderflags_Mode2D3D
+    {
+        e_Mode3D                            = 0x0 << e_Mode2D3DShift,
+        e_Mode2D                            = 0x1 << e_Mode2D3DShift,
+    };
+
+    // Notes:
+    //   Don't change the xxxShift values blindly as they affect the rendering output
+    //   that is two primitives have to be rendered after 3d primitives, alpha blended
+    //   geometry have to be rendered after opaque ones, etc.
+    //   This also applies to the individual flags in EAuxGeomPublicRenderflagBitMasks!
+    // Remarks:
+    //   Bits 0 - 22 are currently reserved for prim type and per draw call render parameters (point size, etc.)
+    //   Check RenderAuxGeom.h in ../RenderDll/Common
+    // See also:
+    //   EAuxGeomPublicRenderflagBitMasks
+    enum EAuxGeomPublicRenderflags_AlphaBlendMode
+    {
+        e_AlphaNone                     = 0x0 << e_AlphaBlendingShift,
+        e_AlphaAdditive             = 0x1 << e_AlphaBlendingShift,
+        e_AlphaBlended              = 0x2 << e_AlphaBlendingShift,
+    };
+
+    // Notes:
+    //   Don't change the xxxShift values blindly as they affect the rendering output
+    //   that is two primitives have to be rendered after 3d primitives, alpha blended
+    //   geometry have to be rendered after opaque ones, etc.
+    //   This also applies to the individual flags in EAuxGeomPublicRenderflagBitMasks!
+    // Remarks:
+    //   Bits 0 - 22 are currently reserved for prim type and per draw call render parameters (point size, etc.)
+    //   Check RenderAuxGeom.h in ../RenderDll/Common
+    // See also:
+    //   EAuxGeomPublicRenderflagBitMasks
+    enum EAuxGeomPublicRenderflags_DrawInFrontMode
+    {
+        e_DrawInFrontOff            = 0x0 << e_DrawInFrontShift,
+        e_DrawInFrontOn             = 0x1 << e_DrawInFrontShift,
+    };
+
+    // Notes:
+    //   Don't change the xxxShift values blindly as they affect the rendering output
+    //   that is two primitives have to be rendered after 3d primitives, alpha blended
+    //   geometry have to be rendered after opaque ones, etc.
+    //   This also applies to the individual flags in EAuxGeomPublicRenderflagBitMasks!
+    // Remarks:
+    //   Bits 0 - 22 are currently reserved for prim type and per draw call render parameters (point size, etc.)
+    //   Check RenderAuxGeom.h in ../RenderDll/Common
+    // See also:
+    //   EAuxGeomPublicRenderflagBitMasks
+    enum EAuxGeomPublicRenderflags_FillMode
+    {
+        e_FillModeSolid             = 0x0 << e_FillModeShift,
+        e_FillModeWireframe     = 0x1 << e_FillModeShift,
+        e_FillModePoint             = 0x2 << e_FillModeShift,
+    };
+
+    // Notes:
+    //   Don't change the xxxShift values blindly as they affect the rendering output
+    //   that is two primitives have to be rendered after 3d primitives, alpha blended
+    //   geometry have to be rendered after opaque ones, etc.
+    //   This also applies to the individual flags in EAuxGeomPublicRenderflagBitMasks!
+    // Remarks:
+    //   Bits 0 - 22 are currently reserved for prim type and per draw call render parameters (point size, etc.)
+    //   Check RenderAuxGeom.h in ../RenderDll/Common
+    // See also:
+    //   EAuxGeomPublicRenderflagBitMasks
+    enum EAuxGeomPublicRenderflags_CullMode
+    {
+        e_CullModeNone              = 0x0 << e_CullModeShift,
+        e_CullModeFront             = 0x1 << e_CullModeShift,
+        e_CullModeBack              = 0x2 << e_CullModeShift,
+    };
+
+    // Notes:
+    //   Don't change the xxxShift values blindly as they affect the rendering output
+    //   that is two primitives have to be rendered after 3d primitives, alpha blended
+    //   geometry have to be rendered after opaque ones, etc.
+    //   This also applies to the individual flags in EAuxGeomPublicRenderflagBitMasks!
+    // Remarks:
+    //   Bits 0 - 22 are currently reserved for prim type and per draw call render parameters (point size, etc.)
+    //   Check RenderAuxGeom.h in ../RenderDll/Common
+    // See also:
+    //   EAuxGeomPublicRenderflagBitMasks
+    enum EAuxGeomPublicRenderflags_DepthWrite
+    {
+        e_DepthWriteOn              = 0x0 << e_DepthWriteShift,
+        e_DepthWriteOff             = 0x1 << e_DepthWriteShift,
+    };
+
+    // Notes:
+    //   Don't change the xxxShift values blindly as they affect the rendering output
+    //   that is two primitives have to be rendered after 3d primitives, alpha blended
+    //   geometry have to be rendered after opaque ones, etc.
+    //   This also applies to the individual flags in EAuxGeomPublicRenderflagBitMasks!
+    // Remarks:
+    //   Bits 0 - 22 are currently reserved for prim type and per draw call render parameters (point size, etc.)
+    //   Check RenderAuxGeom.h in ../RenderDll/Common
+    // See also:
+    //   EAuxGeomPublicRenderflagBitMasks
+    enum EAuxGeomPublicRenderflags_DepthTest
+    {
+        e_DepthTestOn                   = 0x0 << e_DepthTestShift,
+        e_DepthTestOff              = 0x1 << e_DepthTestShift,
+    };
+
+
     /// DebugDisplayRequests provides a debug draw api to be used by components and viewport features.
     class DebugDisplayRequests
         : public AZ::EBusTraits
