+++ Gems/PhysX/Debug/Code/Source/SystemComponent.h
@@ -16,7 +16,6 @@
 #include <PhysXDebug/PhysXDebugBus.h>
 #include <PxPhysicsAPI.h>
 
-#include <IRenderAuxGeom.h>
 #include <CryCommon/CrySystemBus.h>
 
 #include <AzFramework/Physics/SystemBus.h>
